Updated `DarwinCentral` for async/await
